Create a 16 x 16 grid
implemented a function that loops twice. In the first loop it creates a
flex div for acomodating the new divs in a row. In the inner loop it creates
divs with a flex property and then appends them into the divRow.
All elements have a border and the number of row and column for a better
understanding of the distribution.